% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portalled_items.R
\name{get_portalled_items}
\alias{get_portalled_items}
\title{Catalogue all the objects in the abyss.}
\usage{
get_portalled_items()
}
\description{
Where you need a specfic object, knowing its name is necessary to conjur it.
}
\details{
Names have power.
}
