\name{GTEx_V4}
\alias{GTEx_V4}

\title{GTEx cis-eQTLs for multiple tissues (based on Version 4)}

\usage{
GTEx <- xRDataLoader(RData.customised='GTEx_V4')
}

\description{
This dataset contains data obtained from GTEx. For each of 13 tissues, cis-eQTLs are stored.
}

\value{
    a list with 13 components/tisses, each with a data frame having 5 columns: "SNP", "Gene_Name", "Beta", "T_Stat" and "P_Val"
}

\keyword{datasets}

\examples{
GTEx <- xRDataLoader(RData.customised='GTEx_V4')
names(GTEx)
str(GTEx)
head(GTEx$Whole_Blood)
}
