% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Conjur an object from the ether}
\usage{
pull(object)
}
\arguments{
\item{object}{Search the abyss for an object that has this symbol's name. In
the abyss not all searches end.}
}
\description{
Oldest objects emerge first.
}
