% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Push an object into the ether}
\usage{
push(object, serialiser = getOption("portal.serialser", default = "rds"))
}
\arguments{
\item{object}{the object to be thrown into the portal. Beware traveller,
somewhere in the abyss a file takes the symbol's name.}

\item{serialiser}{what kind of file the object is backed by. "rds" (default), "qs", or "parquet"}
}
\description{
Push an object into the ether
}
